Extract port-commit workflow into a lazy-loaded module
